option to install extra packages needed by role
New --extra-install option allows installing extra packages
that are needed to test the role.

In some cases missing software could be considered a bug, meaning
the role-under-test is incomplete, but not always.

For example, I was testing a role that was failing when the docker
instance didn't have an openssh client installed, despite '-c local'.
That's not a concern on a real ansible node, which is guaranteed to
have openssh installed.

I could have added openssh client to the mandatory ansible
prerequisites that are automatically installed, but:
* that list already takes long enough to install, without adding
  packages that are only occasionally needed
* the more general mechanism can come in handy in other cases too